Moved wrapped client into src, guarded by wrapped-client feature
