+++ TestModels/SimpleTypes/SimpleBoolean/runtimes/rust/Cargo.toml
@@ -5,6 +5,9 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+wrapped-client = []
+
 [dependencies]
 aws-smithy-runtime = {version = "1.6.0", features = ["client"] }
 aws-smithy-runtime-api = {version = "1.7.0", features = ["client"] }
@@ -13,6 +16,9 @@ dafny_runtime = { path = "../../../../dafny-dependencies/dafny_runtime_rust"}
 dafny_standard_library = { path = "../../../../dafny-dependencies/StandardLibrary/runtimes/rust"}
 simple_boolean_dafny = { path = "./dafny_impl"}
 
-[dev-dependencies.tokio]
+[dev-dependencies]
+simple_boolean = { path = ".", features = ["wrapped-client"] }
+
+[dependencies.tokio]
 version = "1.26.0"
 features = ["full"]
