+++ TestModels/SimpleTypes/SimpleBoolean/runtimes/rust/src/wrapped.rs
@@ -1,0 +1,1 @@
+pub mod client;
